fix(tts): MPS 에서 float16 대신 float32 사용
b13232f2 에서 MPS 갈래에 float16 을 넣었는데 실측이 뒤집었다. M4 맥미니에서
모델 로드와 화자 인식(speakers=['debi'])까지 지나가고 생성 마지막 샘플링에서
죽는다.

    torch.AcceleratorError: probability tensor contains inf, nan or element < 0

float16 의 표현 범위가 좁아 logits 가 넘친 것이다. float32 로는 통과한다
(로드 11.8s, 3.4초 발화 생성 11.5s). bfloat16 은 float32 가 먼저 통과해
확인하지 않았다.

메모리는 두 배가 되지만 데비 경량판 기준 3.4GB 라 16GB 에서 여유가 있다.
마를렌은 원본이 2배라 이 조합이면 7GB 가 되므로 QWEN3_TTS_LAZY 가 필요하다.

+++ run/services/tts/qwen3_tts_api.py
@@ -81,8 +81,9 @@ def init_device():
     if torch.cuda.is_available():
         device, dtype = "cuda:0", torch.bfloat16
     elif torch.backends.mps.is_available():
-        # MPS 는 bfloat16 커버리지가 얇다. float16 이라야 커널이 다 붙는다.
-        device, dtype = "mps", torch.float16
+        # float16 은 샘플링 단계에서 확률 텐서가 inf/nan 으로 넘친다(M4 실측:
+        # "probability tensor contains inf, nan or element < 0"). float32 라야 통과한다.
+        device, dtype = "mps", torch.float32
     else:
         device, dtype = "cpu", torch.float32
